[fix] #48 mv dtos to right position
