+++ src/main/java/com/telepigeon/server/dto/auth/response/JwtTokensDto.java
@@ -1,4 +1,4 @@
-package com.telepigeon.server.dto.auth;
+package com.telepigeon.server.dto.auth.response;
 
 public record JwtTokensDto(String accessToken, String refreshToken) {
     public static JwtTokensDto of(String accessToken, String refreshToken) {
