--- src/main/java/com/telepigeon/server/dto/naverCloud/DocumentDto.java
@@ -1,7 +1,0 @@
-package com.telepigeon.server.dto.naverCloud;
-
-public record DocumentDto(
-        String sentiment,
-        ConfidenceDto confidence
-) {
-}
