--- src/main/java/com/telepigeon/server/dto/naverCloud/NaverCloudDto.java
@@ -1,6 +1,0 @@
-package com.telepigeon.server.dto.naverCloud;
-
-public record NaverCloudDto(
-        DocumentDto document
-) {
-}
